+++ src/ui/about.rs
@@ -43,7 +43,7 @@ impl SimpleComponent for AboutDialog {
             set_version: &APP_VERSION,
 
             set_developers: &[
-                "Nikita Podvirnyy https://github.com/krypt0nn"
+                "Nikita Podvirnyi https://github.com/krypt0nn"
             ],
 
             add_credit_section: (Some("Patch credits"), &[
@@ -99,28 +99,11 @@ impl SimpleComponent for AboutDialog {
 
             set_release_notes_version: &APP_VERSION,
             set_release_notes: &[
-                "<p>Added</p>",
-
-                "<ul>",
-                    "<li>Bundle applications-system-symbolic icon to the app</li>",
-                    "<li>Added force grab cursor option to the gamescope settings</li>",
-                    "<li>Added Thai</li>",
-                    "<li>Added Ukrainian</li>",
-                "</ul>",
-
                 "<p>Changed</p>",
 
                 "<ul>",
-                    "<li>Update wish url</li>",
-                    "<li>Updated dependencies</li>",
-                    "<li>Improved app args parsing</li>",
-                    "<li>Updated locales</li>",
-                "</ul>",
-
-                "<p>Fixed</p>",
-
-                "<ul>",
-                    "<li>Fixed GtkSwitch UI state representation</li>",
+                    "<li>Updated voiceovers sizes</li>",
+                    "<li>Updated game version parsing algorithm</li>",
                 "</ul>"
             ].join("\n"),